esh: Rewrite the \SpecialChar logic
The original system worked nicely if Emacs was set up to use the same
fonts as XeLaTeX; barring that, however, there can and will be
discrepancies between the system's font coverage and the one used in
XeLaTeX, resulting in missing characters.

This new heuristic assumes that everything beyond ASCII 127 should be
rendered in the fallback font.