server: Don't capture traces unless asked to
Backtraces can be huge, and loading SQL-mode actually throws an error,
which is then caught by sql-mode code. Unfortunately, there seems to be
no way to only capture backtraces for errors that bubble back to the
client, so only capture backtraces when asked (makes a 95%
perf. difference for HTML export).