trees: Prototype an interval tree
The (semi-)streaming approach of the previous commit isn't complete: the issue
is that when multiple properties change on the same position, the order in which
the corresponding open or close tags should be generated isn't clear.  Ordering
them by length of the corresponding intervals could be an interesting direction
to explore, but this implementation takes a different route.