tests: Include a tricky (and currently incorrect) example of overlap
There are two issues with this test: spurious blanks, and sub-optimal nesting.
The latter is likely to require a change of algorithm.

Here's a short description of the problem. Suppose there's one interval
a:5-15. Insert b:1-10, which splits a into a1:5-10 and a2:10-15.  Then insert
c:1-8, which splits a1 into a11:5-8 and a12:8-10, and b into b1:1-8 and b2:8-10.
But now b2 fits into a12, and a11 fits in b1.

The trick is that inserting b1 and b2 into a1':5-8 a2':8-15 would not have
required further splits.  In other words, cutting a to accommodate b wasn't
right, since b was going to be cut later on, rendering the original cut useless.